Refactor WebRTC signaling and command structure
Removed the legacy signaling module and reorganized the command structure into distinct modules for legacy and candidate APIs. Introduced a new logger module for improved logging functionality. Updated the main library file to reflect these changes, ensuring a cleaner and more maintainable codebase. Added utility functions for ICE server configuration and candidate management.
