+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod candidate_api;
+pub mod legacy_api;
+pub mod util_api;
