+++ src-tauri/src/lib.rs
@@ -1,6 +1,8 @@
+mod commands;
 mod config;
-mod signaling;
-mod webrtc_peer;
+mod logger;
+mod peer;
+mod utils;
 
 #[tauri::command]
 fn greet(name: &str) -> String {
@@ -13,22 +15,22 @@ pub fn run() {
         .plugin(tauri_plugin_opener::init())
         .invoke_handler(tauri::generate_handler![
             // Legacy API
-            signaling::generate_offer,
-            signaling::accept_offer_and_create_answer,
-            signaling::set_answer,
+            commands::legacy_api::generate_offer,
+            commands::legacy_api::accept_offer_and_create_answer,
+            commands::legacy_api::set_answer,
             // New API with candidates
-            signaling::generate_offer_with_candidates,
-            signaling::accept_offer_with_candidates,
-            signaling::set_answer_with_candidates,
-            signaling::add_ice_candidate,
+            commands::candidate_api::generate_offer_with_candidates,
+            commands::candidate_api::accept_offer_with_candidates,
+            commands::candidate_api::set_answer_with_candidates,
+            peer::ice::add_ice_candidate,
             // Utility functions
-            signaling::send_text,
-            signaling::get_fingerprint,
-            signaling::is_connected,
-            signaling::disconnect,
-            signaling::check_ice_server_availability,
-            signaling::set_ice_servers,
-            signaling::get_ice_servers,
+            commands::util_api::send_text,
+            commands::util_api::get_fingerprint,
+            commands::util_api::is_connected,
+            commands::util_api::disconnect,
+            peer::ice::check_ice_server_availability,
+            peer::connection::set_ice_servers,
+            peer::connection::get_ice_servers,
             greet
         ])
         .run(tauri::generate_context!())
