+++ src-tauri/src/peer/types.rs
@@ -1,0 +1,68 @@
+use base64::{engine::general_purpose, Engine as _};
+use bytes::Bytes;
+use chacha20poly1305::{
+    aead::{Aead, KeyInit, Nonce},
+    ChaCha20Poly1305, Key,
+};
+use flate2::{read::GzDecoder, write::GzEncoder, Compression};
+use hkdf::Hkdf;
+use once_cell::sync::Lazy;
+use rand::Rng;
+use ring::{agreement, rand as ring_rand};
+use serde::{Deserialize, Serialize};
+use sha2::{Digest, Sha256};
+use std::io::{Read, Write};
+use std::sync::{Arc, Mutex};
+use std::time::Duration;
+use tauri::{AppHandle, Emitter};
+use tokio::sync::mpsc;
+use tokio::time::sleep;
+use tokio::time::timeout;
+use webrtc::ice_transport::ice_candidate::RTCIceCandidate;
+use webrtc::ice_transport::ice_candidate::RTCIceCandidateInit;
+use webrtc::peer_connection::policy::bundle_policy::RTCBundlePolicy;
+use webrtc::peer_connection::policy::rtcp_mux_policy::RTCRtcpMuxPolicy;
+use webrtc::{
+    api::APIBuilder,
+    data_channel::{data_channel_init::RTCDataChannelInit, RTCDataChannel},
+    ice_transport::{ice_gatherer_state::RTCIceGathererState, ice_server::RTCIceServer},
+    peer_connection::{
+        configuration::RTCConfiguration, peer_connection_state::RTCPeerConnectionState,
+        sdp::session_description::RTCSessionDescription, RTCPeerConnection,
+    },
+};
+use zeroize::{Zeroize, ZeroizeOnDrop};
+
+/// Полезная нагрузка SDP с метаданными
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct SdpPayload {
+    pub sdp: RTCSessionDescription,
+    pub id: String,
+    pub ts: i64,
+}
+
+/// ICE кандидат для WebRTC соединения
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct IceCandidate {
+    pub candidate: String,
+    pub sdp_mid: Option<String>,
+    pub sdp_mline_index: Option<u16>,
+    pub connection_id: String, // ID соединения для сопоставления
+}
+
+/// Полный пакет соединения с SDP и кандидатами
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct ConnectionBundle {
+    pub sdp_payload: SdpPayload,
+    pub ice_candidates: Vec<IceCandidate>,
+}
+
+/// Конфигурация ICE сервера
+#[derive(Serialize, Deserialize, Debug, Clone)]
+pub struct ServerConfig {
+    pub id: String,
+    pub r#type: String, // 'stun' or 'turn'
+    pub url: String,
+    pub username: Option<String>,
+    pub credential: Option<String>,
+}
